realizando request do arquivo... ao api...
